ai-service/requirements.txt
# AI Service Dependencies
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
python-dotenv==1.0.1

# LangChain & LangGraph for AI Agents
langchain==0.3.13
langchain-openai==0.3.0
langgraph==0.2.60

# OpenAI
openai==1.58.1

# Profanity Detection (Multi-language)
better-profanity==0.7.0
profanity-check==1.0.3

# HTTP Client for backend communication
httpx==0.28.1
aiohttp==3.11.11

# Data Processing
numpy==2.2.1
pandas==2.2.3

# Caching
cachetools==5.5.0

# CORS
python-multipart==0.0.19
